Fixes the problems with client hanging during server restarts on heavy loads. Fixes a memory access issue when re-throwing an exception, correctly clears the call promise maps during connection close. Adds LoadTests while server retarting and solves the problems that the tests creates.

+++ hazelcast/src/hazelcast/client/proxy/IQueueImpl.cpp
@@ -105,7 +105,7 @@ namespace hazelcast {
                 return invokeAndGetResult<bool, protocol::codec::QueueContainsCodec::ResponseParameters>(request, partitionId);
             }
 
-            std::vector<serialization::pimpl::Data> IQueueImpl::drainTo(int maxElements) {
+            std::vector<serialization::pimpl::Data> IQueueImpl::drainTo(size_t maxElements) {
                 std::auto_ptr<protocol::ClientMessage> request =
                         protocol::codec::QueueDrainToMaxSizeCodec::RequestParameters::encode(getName(), maxElements);
